# Add argument to suppress whitespace