--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -27,6 +27,7 @@ Collate:
     'html_dependency.R'
     'html_escape.R'
     'html_print.R'
+    'shim.R'
     'utils.R'
     'tags.R'
     'template.R'