… NAMESPACE
… man/builder.Rd
… man/tag.Rd
… tests/testthat/test-tags.r
… tests/testthat/test-textwriter.r
… tests/testthat/test-whitespace.r