- [x] Need to make `.noWS` a formal argument instead of an undocumented attribute
- [x] Need to document the enumerated values for .noWS
- [x] Validate .noWS arguments to ensure that nothing's misspelled
- [x] Update NEWS, bump version?
- [x] Why is .noWS whitespace-delimted? Shouldn't it just be a regular character vector?