@trestletech 2019-07-05
I did some quick performance tests and it looks like this slows things down quite a bit. I haven't profiled to see where the time is going, but I'm seeing 60-100% extra time taken relative to master:

```
size <- 10000
system.time(l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x, .noWS="outside before")})))
system.time(s <- as.character(l))

### .noWS branch
#> size <- 10000
#> system.time(l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)})))
#user  system elapsed 
#0.253   0.001   0.254 
#> system.time(s <- as.character(l))
#user  system elapsed 
#2.423   0.169   2.596

### .noWS branch with `.noWS="outside before"`
# > size <- 10000
# > system.time(l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x, .noWS="outside before")})))
# user  system elapsed 
# 0.250   0.001   0.251 
# > system.time(s <- as.character(l))
# user  system elapsed 
# 2.774   0.179   2.957 

### master
#> size <- 10000
#> system.time(l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)})))
#user  system elapsed 
#0.246   0.005   0.252 
#> system.time(s <- as.character(l))
#user  system elapsed 
#1.574   0.014   1.592
```