@trestletech 2019-07-09
After a bit of Travis thrashing followed by a force-push-fest to clean up the commit mess... ready for rereview, @jcheng5!

The one outstanding concern for me is the performance issue described in the comment above. I'm wondering if we should hold this PR back until we take on some of the C++ performance improvements. I think I could start on that work whether or not this is on master so long as we're not anticipating any serious conflicts to emerge in the coming couple of weeks.