@trestletech 2019-07-10
I'm doing some experimentation on performance [here](https://github.com/rstudio/htmltools/commits/jeff/feature/stringbuilder). Raw performance notes from running this code:

```
size <- 1000
l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)}))
bench::mark(min_time=5,{
  s <- as.character(l)
})
```

By the end, it looks like we have an approach that will get us back to parity with master if we clean up that branch.

### master

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory time 
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list> <lis>
1 { s <- as.character(l) } 152ms  173ms      5.71     430KB     7.38    58    75      10.2s <chr … <df[,… <bch…
# … with 1 more variable: gc <list>
```

### binary connection-based

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 213ms  239ms      4.20     508KB     8.80    21    44
# … with 5 more variables: total_time <bch:tm>, result <list
```

### buffer-based instead of connection (b09ee)

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 295ms  326ms      3.05     129MB     11.3    16    59
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```

### buffer-based with periodic string collection (345eff0)

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 220ms  235ms      4.28    4.39MB     9.15    22    47
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```

### buffer-based with string collection and double-assign to avoid copies

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 200ms  221ms      4.51    40.1MB     9.42    23    48
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```

### Same with portable=FALSE (aa8c3ad)

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
1 { s <- as.character(l) } 193ms  210ms      4.74     293KB     10.1    24    51      5.07s
# … with 4 more variables: result <list>, memory <list>, time <list>, gc <list>
```

### Same after eliminating TextConnection and just using WS directly (de642607)

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory time 
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list> <lis>
1 { s <- as.character(l) } 1.71s  1.77s     0.566    9.85MB     9.34     6    99      10.6s <chr … <df[,… <bch…
# … with 1 more variable: gc <list>

### after removing R6 and using a closure directly (5e5bab0c)

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory time 
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list> <lis>
1 { s <- as.character(l) } 142ms  166ms      5.74     271KB     7.72    58    78      10.1s <chr … <df[,… <bch…
# … with 1 more variable: gc <list>
```