@amrrs 2020-01-13
@jcheng5 @wch could it be used with `htmlTemplate()`. It's still adding linebreaks