@jcheng5 2020-01-13
@amrrs Oh, that's a good point. I think we'd actually need to add this feature to `HTML()` as well. This code:

```
div(HTML("a"), HTML("b"), .noWS = c("after-begin", "before-end"))
```

renders like:

```
<div>a
  b</div>
```

There's not currently a way AFAIK to remove the whitespace between `a` and `b`, and that's basically what we'd need to make `htmlTemplate` eat the whitespace.