@jcheng5 2020-01-15
@jcubic That's what I was referring to in my previous comment, sorry it wasn't clear. `htmlTemplate` is essentially implemented by stringing together `HTML()` segments, as soon as we support whitespace eating around those segments then `{{ ... }}` can be fixed.